+++ frontend/src/router/routes/modules/bpm.ts
@@ -51,22 +51,16 @@ const bpm: AppRouteModule = {
       meta: { title: '我发起的' },
     },
     {
-      path: 'approve',
-      name: 'BpmApprove',
-      component: () => import('/@/views/bpm/approve/index.vue'),
-      meta: { title: '审批页面' },
-    },
-    {
       path: 'done',
       name: 'bpm-done',
       component: () => import('/@/views/bpm/done/index.vue'),
       meta: { title: '已办任务' },
     },
     {
-      path: 'my',
-      name: 'bpm-my',
-      component: () => import('/@/views/bpm/bpm-my-initiated/index.vue'),
-      meta: { title: '我发起的' },
+      path: 'approve',
+      name: 'BpmApprove',
+      component: () => import('/@/views/bpm/approve/index.vue'),
+      meta: { title: '审批页面' },
     },
     {
       path: 'process/view',